fix: Fix Init command
On a workspace without Tauri (e.g. a workspace generated with
`pnpm create vite .`), the `Tauri: Init` command fails with
error message `Tauri project not found`.

That's because of a regression in #242:
https://github.com/tauri-apps/tauri-vscode/pull/242/files#diff-04bba6a35cad1c794cbbe677678a51de13441b7a6ee8592b7b50be1f05c6f626L199

The Init command should search a directory with NO subdirectory
named `src-tauri`, because Tauri is not installed yet.